steggy/steggy/requirements.txt
colorama
pycryptodome
exifread
random2
Pillow
